Move coll-obj-to-arr util into separate file
