+++ workers/loc.api/helpers/utils/index.js
@@ -1,5 +1,6 @@
 'use strict'
 
+const collObjToArr = require('./coll-obj-to-arr')
 const getDateString = require('./get-date-string')
 const isNotSyncRequired = require('./is-not-sync-required')
 const pickLowerObjectsNumbers = require('./pick-lower-objects-numbers')
@@ -11,6 +12,7 @@ const pushLargeArr = require('./push-large-arr')
 const orderBy = require('./order-by')
 
 module.exports = {
+  collObjToArr,
   getDateString,
   isNotSyncRequired,
   pickLowerObjectsNumbers,
